+++ src/core/drd_config.c
@@ -91,6 +91,7 @@ drd_config_init(DrdConfig *self)
     self->encoding.h264_framerate = DRD_H264_DEFAULT_FRAMERATE;
     self->encoding.h264_qp = DRD_H264_DEFAULT_QP;
     self->encoding.h264_hw_accel = DRD_H264_DEFAULT_HW_ACCEL;
+    self->encoding.h264_vm_support = DRD_H264_DEFAULT_VM_SUPPORT;
     self->encoding.gfx_large_change_threshold = DRD_GFX_DEFAULT_LARGE_CHANGE_THRESHOLD;
     self->encoding.gfx_progressive_refresh_interval = DRD_GFX_DEFAULT_PROGRESSIVE_REFRESH_INTERVAL;
     self->encoding.gfx_progressive_refresh_timeout_ms = DRD_GFX_DEFAULT_PROGRESSIVE_REFRESH_TIMEOUT_MS;
@@ -480,6 +481,18 @@ drd_config_load_from_key_file(DrdConfig *self, GKeyFile *keyfile, GError **error
             return FALSE;
         }
         self->encoding.h264_hw_accel = value;
+    }
+
+    if (g_key_file_has_key(keyfile, "encoding", "h264_vm_support", NULL))
+    {
+        g_autofree gchar *vm_support =
+                g_key_file_get_string(keyfile, "encoding", "h264_vm_support", NULL);
+        gboolean value = DRD_H264_DEFAULT_VM_SUPPORT;
+        if (!drd_config_parse_bool(vm_support, &value, error))
+        {
+            return FALSE;
+        }
+        self->encoding.h264_vm_support = value;
     }
 
     if (g_key_file_has_key(keyfile, "encoding", "gfx_large_change_threshold", NULL))